add chai, add more tests
