+++ mdx/emit.ts
@@ -1,15 +1,13 @@
 import { readdir, rm, stat } from "fs-extra";
 import path from "path";
-import {
-  convertPackageToMdx,
-  createDirectoryIfNotExist,
-  emitCategoryMetadata,
-  emitMdx,
-  emitMessagesJson,
-  getServiceString,
-} from "./lib/doc-to-mdx";
-import { convertProtoToRecord } from "./lib/record";
+import { emitMdx } from "./lib/mdx";
+import { convertProtoToRecord } from "./lib/utils/record";
 import { PackageData, ProtoMessage } from "./lib/types";
+import { emitMessagesJson } from "./lib/json";
+import { emitCategoryMetadata } from "./lib/metadata";
+import { readPackageData } from "./lib/packages";
+import { getServiceString } from "./lib/services";
+import { createDirectoryIfNotExist } from "./lib/utils/filesystem";
 
 // These are meant only to be ran from the Makefile to take effect of the PWD environment variable.
 // This is because, after being compiled to `.js` files, they go into a deeper nested directories,
@@ -165,8 +163,9 @@ async function recursivelyReadDirectory({
 
     // If file, read its contents.
     if (entry.isFile()) {
-      const { packageData: packages, rawProtoMessages } =
-        await convertPackageToMdx(pathToEntry);
+      const { packageData: packages, rawProtoMessages } = await readPackageData(
+        pathToEntry
+      );
 
       allPackages.push(...packages);
       allProtoMessages.push(...rawProtoMessages);
