+++ mdx/lib/test-resources/booking-services.mdx
@@ -29,10 +29,6 @@ returned.
   Booking
 </RpcMethodText>
 
-<Definition>
-
-
-
 ```protosaurus--booking.v1.Booking
 message Booking {
   // ID of booked vehicle.
@@ -56,13 +52,7 @@ message Booking {
 }
 ```
 
-</Definition>
-
 <RpcMethodText type="response" isStream={false}>BookingStatus</RpcMethodText>
-
-<Definition>
-
-
 
 ```protosaurus--booking.v1.BookingStatus
 message BookingStatus {
@@ -74,8 +64,6 @@ message BookingStatus {
   Duration duration = 3;
 }
 ```
-
-</Definition>
 
 </RpcDefinitionDescription>
 
@@ -101,10 +89,6 @@ Used to subscribe to updates of the BookingStatus.
   BookingStatusID
 </RpcMethodText>
 
-<Definition>
-
-
-
 ```protosaurus--booking.v1.BookingStatusID
 message BookingStatusID {
   // Unique booking status ID.
@@ -112,13 +96,7 @@ message BookingStatusID {
 }
 ```
 
-</Definition>
-
 <RpcMethodText type="response" isStream={true}>BookingStatus</RpcMethodText>
-
-<Definition>
-
-
 
 ```protosaurus--booking.v1.BookingStatus
 message BookingStatus {
@@ -130,8 +108,6 @@ message BookingStatus {
   Duration duration = 3;
 }
 ```
-
-</Definition>
 
 </RpcDefinitionDescription>
