+++ mdx/lib/test-resources/location-messages.mdx
@@ -15,6 +15,7 @@ message Location {
   // Longitude of the location.
   int32 longitude = 2;
   // Address (human-friendly string) of the location.
+  // This line tests a <html> tag comment.
   string address = 3;
 
   message InnerMessage {
