+++ mdx/lib/utils/filesystem.ts
@@ -1,0 +1,10 @@
+import { mkdirp, stat } from "fs-extra";
+
+export async function createDirectoryIfNotExist(directory: string) {
+  try {
+    await stat(directory);
+  } catch (err) {
+    // Not found.
+    await mkdirp(directory);
+  }
+}
