+++ mdx/package.json
@@ -3,9 +3,11 @@
   "version": "0.0.1",
   "license": "MIT",
   "devDependencies": {
+    "@types/chai": "4.3.0",
     "@types/fs-extra": "9.0.13",
     "@types/mocha": "9.1.0",
     "@types/node": "17.0.12",
+    "chai": "4.3.6",
     "mocha": "9.2.0",
     "ts-node": "10.4.0",
     "typescript": "4.5.5"
